Split all models and crud into individual files for cleanliness. Updated all other necessary imports. Might not be fully done, testing now.
