+++ backend/app/api/routes/login.py
@@ -5,12 +5,13 @@ from fastapi import APIRouter, Depends, HTTPException
 from fastapi.responses import HTMLResponse
 from fastapi.security import OAuth2PasswordRequestForm
 
-from app.crud import crud
+from app.crud import users as crud_users
 from app.api.deps import CurrentUser, SessionDep, get_current_active_superuser
 from app.core import security
 from app.core.config import settings
 from app.core.security import get_password_hash
-from app.models.models import Message, NewPassword, Token, UserPublic
+from app.models.utils import Message, NewPassword, Token
+from app.models.users import UserPublic
 from app.utils import (
     generate_password_reset_token,
     generate_reset_password_email,
@@ -28,7 +29,7 @@ def login_access_token(
     """
     OAuth2 compatible token login, get an access token for future requests
     """
-    user = crud.authenticate(
+    user = crud_users.authenticate(
         session=session, email=form_data.username, password=form_data.password
     )
     if not user:
@@ -56,7 +57,7 @@ def recover_password(email: str, session: SessionDep) -> Message:
     """
     Password Recovery
     """
-    user = crud.get_user_by_email(session=session, email=email)
+    user = crud_users.get_user_by_email(session=session, email=email)
 
     if not user:
         raise HTTPException(
@@ -83,7 +84,7 @@ def reset_password(session: SessionDep, body: NewPassword) -> Message:
     email = verify_password_reset_token(token=body.token)
     if not email:
         raise HTTPException(status_code=400, detail="Invalid token")
-    user = crud.get_user_by_email(session=session, email=email)
+    user = crud_users.get_user_by_email(session=session, email=email)
     if not user:
         raise HTTPException(
             status_code=404,
@@ -107,7 +108,7 @@ def recover_password_html_content(email: str, session: SessionDep) -> Any:
     """
     HTML Content for Password Recovery
     """
-    user = crud.get_user_by_email(session=session, email=email)
+    user = crud_users.get_user_by_email(session=session, email=email)
 
     if not user:
         raise HTTPException(
