+++ backend/app/api/routes/private.py
@@ -5,7 +5,7 @@ from pydantic import BaseModel
 
 from app.api.deps import SessionDep
 from app.core.security import get_password_hash
-from app.models.models import (
+from app.models.users import (
     User,
     UserPublic,
 )
