+++ backend/app/core/db.py
@@ -2,7 +2,7 @@ from sqlmodel import Session, create_engine, select
 
 from app.crud import crud
 from app.core.config import settings
-from app.models.models import User, UserCreate
+from app.models.users import User, UserCreate
 
 engine = create_engine(str(settings.SQLALCHEMY_DATABASE_URI))
